Fixed testing for the driving fault card
