+++ src/app/pages/office/components/driving-faults/driving-faults.component.html
@@ -17,8 +17,7 @@
             </ion-col>
             <ion-col size="1"></ion-col>
             <ion-col class="component-label ion-align-self-center">
-                <label class="fault-label"
-                       id="driving-faults-label-{{drivingFault.competencyIdentifier}}">{{drivingFault.competencyDisplayName}}</label>
+                <label class="fault-label" id="driving-faults-label-{{drivingFault.competencyIdentifier}}">{{drivingFault.competencyDisplayName}}</label>
             </ion-col>
         </ion-row>
     </ion-card-content>
